+++ src/cli/list/uss-files/UssFiles.definition.ts
@@ -16,14 +16,19 @@ export const ListUssFilesDefinition: ICommandDefinition = {
     type: "command",
     name: "uss-files", aliases: ["uss"],
     summary: "List USS files in a directory",
-    description: "List USS files and subdirectories in a directory.\n" +
-    "The following values can be used with the --response-format-filter (--rff) argument to" +
-    "display more data from the data sets:" +
-    "name, size, owner, group, and permissions",
+    description: "List USS files and subdirectories in a directory. Optional file name pattern like " +
+    "\"prefix*\", \"*suffix\", or \"prefix*suffix\" can be specified at the end of directory. " +
+    "See EXAMPLES section.\n\n" +
+    "The following values can be used with the --response-format-filter (--rff) argument to " +
+    "display more data from the data sets: " +
+    "name, size, owner, group, and permissions.",
     examples: [
         {
             description: "List USS files in the directory \"/u/users/ibmuser/\"",
             options: "\"/u/users/ibmuser\""
+        }, {
+            description: "List USS files with suffix of \".txt\" in the directory \"/u/users/ibmuser/\"",
+            options: "\"/u/users/ibmuser/*.txt\""
         }, {
             description: "List USS files in the directory \"/u/users/ibmuser/\" and show only the file name",
             options: "\"/u/users/ibmuser/\" --rff name"
@@ -31,7 +36,7 @@ export const ListUssFilesDefinition: ICommandDefinition = {
     ],
     positionals: [{
         name: "directory",
-        description: "The USS directory to list files in",
+        description: "The USS directory to list files in, or the directory with file name pattern",
         type: "string",
         required: true
     }],